requirements.txt
langchain_community
langchain_openai
python-dotenv
xata